+++ packages/xbell/src/types/test.ts
@@ -2,50 +2,12 @@ import type { expell } from 'expell';
 import type { XBellTaskConfig, XBellRuntimeOptions, XBellProject } from './config';
 import type { Expect } from '../worker/expect/expect';
 import type { XBellPage } from './page';
-
-export type XBellError = { name: string; message: string; stack?: string };
-
-export type XBellTestCaseStatus =
-  | 'successed'
-  | 'failed'
-  | 'running'
-  | 'waiting';
-
-export interface XBellTestCaseRecord {
-  type: 'case';
-  uuid: string;
-  filename: string;
-  groupDescription?: string;
-  caseDescription: string;
-  status: XBellTestCaseStatus;
-  error?: XBellError;
-  coverage?: any;
-}
-
-export interface XBellTestGroupRecord {
-  type: 'group';
-  filename: string;
-  uuid: string;
-  groupDescription: string;
-  cases: XBellTestTaskRecord[];
-}
-
-export type XBellTestTaskRecord =
-  | XBellTestGroupRecord
-  | XBellTestCaseRecord;
-
-
-export interface XBellTestFileRecord {
-  filename: string;
-  tasks: XBellTestTaskRecord[];
-  logs: XBellWorkerLog[];
-  error?: XBellError;
-}
-
-export interface XBellWorkerLog {
-  type: 'stdout' | 'stderr';
-  content: string;
-}
+import type {
+  XBellTestCaseStatus,
+  XBellError,
+  XBellTestFileRecord,
+  XBellWorkerLog,
+} from './record';
 
 export interface XBellTestFile {
   filename: string;
